jd/requirements.txt
Django==5.1.2
psycopg2-binary==2.9.10
celery
requests
django-redis
flower
unidecode
pytils
djoser
djangorestframework
djangorestframework-simplejwt
colorama
django-filter
